pyproject.toml
[project]
name = "fea-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bottleneck>=1.6.0",
    "gmsh>=4.15.0",
    "matplotlib>=3.10.8",
    "mpi4py>=4.1.1",
    # "numpy>=2.4.1",
    "numexpr>=2.14.1",
    "numpy<2.0",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
    "torch-utils>=0.1.2",
]



[tool.pytest.ini_options]
# This tells pytest where the source code is so you don't need PYTHONPATH
pythonpath = ["src", "."]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.uv]
dev-dependencies = ["pytest>=8.0.0",]
